add DecodeTag, fix DecodeValue signature
Add this function to the cbor-decode package:

DecodeTag([]byte) (uint64, []byte, error)

DecodeTag decodes a CBOR tag from a byte slice.  It returns the tag
number, the tag data, and an error.  If the byte slice does not contain
a tag, it returns 0, nil, nil.

Fix the signature of DecodeValue to look like this:

DecodeValue([]byte, interface{}) error

In other words, we will pass a byte slice and a pointer to a value to
decode into, rather than using generics.