remove DecodeContent, add DecodeRaw
Remove the DecodeContent function from the cbor-decode package because
it is redundant with DecodeInto.

Add a DecodeRaw function to the cbor-decode package:

DecodeRaw([]byte, interface{}) error

DecodeRaw decodes a CBOR value from a byte slice into the given value.
It does not recursively decode the value.  It is useful for decoding
the payload of a CBOR that was returned from DecodeTag.