describe a WASM version of the hello demo
Write a README.md file that lists the files needed for a version of
the hello demo that runs in a web browser using WASM. The file list
should include all files needed to build and run the demo. The file
list should consist of the name of each file and a one-line description.