add Sally to discussion script
Sally is trying to write user stories for the system.  She has a list
of requirements that she's trying to turn into user stories, but she's
having trouble.  She's not sure how to write user stories that are
both specific and general enough to be useful.  She's also not sure
how to write user stories that are testable.  She's looking for help
with this.  Write Sally's part of the discussion script, gleaning
requirements from the available documents and asking questions where
necessary.