publish an IPNS record
In a subroutine, publish an IPNS record for the root CID of the
three-node DAG.  The record should be signed with the private key
named "three-node-dag-test".  If the private key does not exist,
create it.

Refer to /home/stevegt/lab/ipfs/kubo/client/rpc/name.go for how to
publish an IPNS record.