ipld-path-cross-explore-patch: fix errors
What is PopSegment supposed to do? If it exists in the IPLD libraries,
where is it defined? If it doesn't exist, what is the correct way to
do what the code is trying to do?

vet: ./ipldpath.go:143:31: path.PopSegment undefined (type datamodel.Path has no field or method PopSegment)