wire: WIP test case questions
Draft a detailed set of test-case questions, in English, to be used to
evaluate the wire.md document for consistency and compliance with the
requirements for the PromiseGrid wire protocol.  Include wire-chat.md and
web sources as references.  The test cases should be in the form of
questions that can be answered with a yes or no; a yes answer means that
the document is compliant with the requirement.

PromiseGrid web resources include:
- github.com/promisegrid/promisegrid/README.md
- github.com/promisegrid/paper-ism/README.adoc
- github.com/promisegrid/grid-cli/v2/doc/