wire: use a proper CBOR tag as message header
Refactor the message header. The message header should start with a
CBOR tag.  The tag number should be the ascii characters 'grid'
translated to single integer.  The tag content should be a a 2-element
CBOR array, with the first element being the protocol CID and the
second element being the payload.  The protocol CID should be a CIDv1
of a SHA256 hash.