add gossipsub
Add a gossipsub demo to node-boxo-bitswap-plus.  The demo should
simply send "hello world" to the peer; the peer should receive the
message and send a response back to the sender.  The sender should print
the response to the console.

Move the gossipsub and bitswap demos from main() into their own
functions.