add -t flag
When the -d flag is provided, the code will connect to the peer
provided as the flag's argument.  Now that we have a working
DHT implementation, we can use the DHT to find the peer.  So let's add
a -t flag that is a boolean that, if provided, will cause the same
code that runs with -d to run, but will not proactively connect to a
particular peer, but will instead rely on the DHT.

Update the test cases accordingly.