In routing.md, brainstorm how PromiseGrid might do multi-hop routing.
Add a protocol in which the goal is to build a directed hypergraph.
Each message contains the CID of one or more previous nodes,
and also contains the CID of one or more new nodes that are
added to the hypergraph.  The hypergraph is a directed acyclic graph
that represents a set of worldlines.

The routing method is to forward the message to interested agents,
where "interested" means that the agent has previously declared that
it holds a copy of the CID of an ancestor of the nodes in the message,
where "ancestor" means one of the CIDs in the message's routing
information array, or one of their ancestors in the hypergraph.

Give an example of how this might work in practice, using an Alice,
Bob, Carol, and Dave scenario.